Separated the formDraft utils
